% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{compute_pvalue_from_tscore}
\alias{compute_pvalue_from_tscore}
\title{score are assume to follow student distibution with df degre of freedom}
\usage{
compute_pvalue_from_tscore(score, df)
}
\description{
score are assume to follow student distibution with df degre of freedom
}
