% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{compute_pvalue_from_zscore}
\alias{compute_pvalue_from_zscore}
\title{score are assume to follow normal distibution}
\usage{
compute_pvalue_from_zscore(score, mean = 0, sd = 1)
}
\description{
score are assume to follow normal distibution
}
