% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example.data}
\alias{example.data}
\title{Genetic and phenotypic data for Arabidopsis thaliana}
\format{A list with 4 arguments: genotype, phenotype, causal.set, chrpos}
\usage{
data(example.data)
}
\description{
A dataset containing SNP frequency and simulated phenotypic data for 170 plant accessions.
The variables are as follows:
}
\details{
\itemize{
\item genotype: binary (0 or 1) SNP frequency for 170 individuals (26943 SNPs).
\item phenotype: simulated phenotypic data for 170 individuals.
\item causal.set: set of indices for causal SNPs.
\item chrpos: genetic map including chromosome position of each SNP.
}
}
\keyword{datasets}
