% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfmm.R
\name{hypothesis_test_lfmm}
\alias{hypothesis_test_lfmm}
\title{Hypothesis testing of association of Y with X with correction by latent factor.}
\usage{
hypothesis_test_lfmm(Y, X, lfmm, calibrate = TRUE)
}
\arguments{
\item{Y}{Explained variables matrix. Each column is a variable.}

\item{X}{Explaining variables matrix. Each column is a variable.}

\item{lfmm}{Object returned by lasso_lfmm or ridge_lfmm}

\item{calibrate}{If TRUE pvalue are calibrated by computing mean and MAD of
zscore.}
}
\value{
A list with pvalue, zscore and effect.
}
\description{
This function compute the pvalue of the association test of each column of Y
with X. The hypothesis testing take into account latent variables computed
by lasso_lfmm or ridge_lfmm.
}
\author{
cayek
}
