% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis-testing.R
\name{hypothesis_testing_lm}
\alias{hypothesis_testing_lm}
\title{Hypothesis testing with lm}
\usage{
hypothesis_testing_lm(dat, X, lambda)
}
\description{
linear model:
Y = X B^T + E
}
\author{
cayek, francoio
}
