% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfmm.R
\name{iterate_testing}
\alias{iterate_testing}
\title{Recursive tests with latent factor mixed models}
\usage{
iterate_testing(Y, X, K, niter = 20, scale = FALSE, lambda = 1e-04)
}
\arguments{
\item{Y}{a response variable matrix with n rows and p columns.
Each column is a response variable (numeric).}

\item{X}{an explanatory variable matrix with n rows and d = 1 column (eg. phenotype).}

\item{K}{an integer for the number of latent factors in the regression model.}

\item{niter}{an integer value for the number of recursive tests.}

\item{scale}{a local value, \code{TRUE} if the explanatory variable need to be scaled
(recommended option).}

\item{lambda}{a numeric value for the regularization parameter.}
}
\value{
a list with the following attributes:
\itemize{
\item candidates a vector of niter candidate variables (columns of Y),
\item log.p a vector of uncorrected log p-values for checking (not trustable for testing).
}
}
\description{
This function tests for association between each column of the response matrix, Y,
and the explanatory variables, X, by recursively including the top hits in the set
of explanatory variables. The recursive tests are based on LFMMs fitted with ridge
penalty.
}
\details{
The response variable matrix Y and the explanatory variable are centered.
}
\examples{
library(lfmm)
data("example.data")
Y <- example.data$genotype
X <- example.data$phenotype #scaled variable

obj <- iterate_testing(Y, X, K = 6, niter = 22, scale = TRUE)

#perfect hits for each causal SNPs (1-20)
obj$candidate \%in\% example.data$causal.set

#check candidates at distance 20  (about 10kb)
theta <- 20
#number of hits for each causal SNPs (1-20)
 large.list <- as.numeric(
  apply(sapply(obj$candidate, 
  function(x) abs(x - example.data$causal.set) < theta), 
  2, 
  which))
table(large.list)

# Plot log P
plot(obj$log.p, xlab = "Iteration")
}
\author{
cayek, francoio
}
