% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lassoLFMM.R
\name{lassoLFMM}
\alias{lassoLFMM}
\title{Lasso LFMM}
\usage{
lassoLFMM(K, nozero.prop = 0.1, lambda.K = 100, lambda.eps = 0.001)
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
cayek
}
