% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix-factorizationR.R
\docType{package}
\name{lfmm}
\alias{lfmm}
\alias{lfmm-package}
\title{R package : Fast and Accurate statistical methods for adjusting confounding factors
in association studies.}
\description{
Implements statistical methods for adjusting confounding factors
in association studies.
}
\references{
Caye, K., B. Jumentier, J. Lepeule, and O. François, 2019  LFMM 2: fast and accurate inference of gene-environment associations in genome-widestudies. Mol. Biol. Evol. 36: 852–860.https://doi.org/10.1093/molbev/msz008

B. Jumentier, Caye, K., J. Lepeule, and O. François, 2019 Sparse latent factor regression models for genome-wide and epigenome-wide association studies (in prep)
}
