% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix-factorizationR.R
\name{lfmm_fit_knowing_loadings}
\alias{lfmm_fit_knowing_loadings}
\title{Fit the model when latent factor loadings are known}
\usage{
lfmm_fit_knowing_loadings(m, dat, ...)
}
\description{
Fit the model when latent factor loadings are known
}
