% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgeLFMM.R
\name{lfmm_fit_knowing_loadings.ridgeLFMM}
\alias{lfmm_fit_knowing_loadings.ridgeLFMM}
\title{Fit assuming V and B}
\usage{
\method{lfmm_fit_knowing_loadings}{ridgeLFMM}(m, dat)
}
\description{
Fit assuming V and B
}
