% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfmm.R
\name{lfmm_lasso}
\alias{lfmm_lasso}
\title{LFMM \eqn{L_1} regularized estimator}
\usage{
lfmm_lasso(Y, X, K, nozero.prop = 0.1, lambda.num = 100,
  lambda.min.ratio = 0.01, lambda = NULL, it.max = 100,
  relative.err.epsilon = 1e-06)
}
\arguments{
\item{Y}{Explained variables matrix. Each column is a variable.}

\item{X}{Explaining variables matrix. Each column is a variable.}

\item{K}{Number of latent factor.}

\item{nozero.prop}{The proportion of line of B expected to be different from
zero.}

\item{lambda.num}{The number of 'lambda' values.}

\item{lambda.min.ratio}{Smallest value for \code{lambda}, as a fraction of
\code{lambda.max}, the (data derived) entry value (i.e. the smallest value for
which all coefficients are zero).}

\item{lambda}{Smallest value for \code{lambda}, as a fraction of 'lambda.max',
the (data derived) entry value (i.e. the smallest value for which all
coefficients are zero).}

\item{it.max}{The number of iteration of the algorithm.}

\item{relative.err.epsilon}{The relative error used to determine if the
algorithm converged.}
}
\value{
A lfmm class object with following attribute: - U the latent
variable score matrix. - V the latent variable axes matrix. - B the effect
size matrix.
}
\description{
This function compute the $L_1$ regularized least squares estimator of LFMM.
The algorithm optimize the loss function
}
\details{
\deqn{ Llasso(U, V, B) =
frac{1}{2} ||Y - U V^{T} - X B^T||_{F}^2 + \frac{\lambda}{2}
\norm{B}^{2}_{2} }.
}
\examples{
library(lfmm)

## sample data
K <- 3
dat <- lfmm_sampler(n = 100, p = 1000, K = K,
                    outlier.prop = 0.1,
                    cs = c(0.6),
                    sigma = 0.2,
                    B.sd = 1.0,
                    U.sd = 1.0,
                    V.sd = 1.0)
## run lfmm
lfmm.res <- lfmm_lasso(Y = dat$Y, X = dat$X, K = 3, nozero.prop= 0.2)

## plot size effect matrix
id <- seq_along(lfmm.res$B)
cols <- c('red', 'green')[as.numeric(id \%in\% dat$outlier) + 1]
plot(id, lfmm.res$B, col = cols)
}
\author{
cayek
}
