% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfmm.R
\name{lfmm_ridge}
\alias{lfmm_ridge}
\title{LFMM least-squares estimates with ridge penalty}
\usage{
lfmm_ridge(Y, X, K, lambda = 1e-05)
}
\arguments{
\item{Y}{a response variable matrix with n rows and p columns.
Each column corresponds to a distinct response variable (e.g., SNP genotype,
gene expression level, beta-normalized methylation profile, etc).
Response variables must be encoded as numeric.}

\item{X}{an explanatory variable matrix with n rows and d columns.
Each column corresponds to a distinct explanatory variable (eg. phenotype).
Explanatory variables must be encoded as numeric.}

\item{K}{an integer for the number of latent factors in the regression model.}

\item{lambda}{a numeric value for the regularization parameter.}
}
\value{
an object of class \code{lfmm} with the following attributes:
\itemize{
\item U the latent variable score matrix with dimensions n x K.
\item V the latent variable axis matrix with dimensions p x K.
\item B the effect size matrix with dimensions p x d.
}
}
\description{
This function computes regularized least squares estimates
for the parameters of latent factor mixed models using a ridge penalty.
}
\details{
The algorithm minimizes the following penalized least-squares criterion
\deqn{ Lridge(U, V, B) = \frac{1}{2} ||Y - U V^{T} - X B^T||_{F}^2 
+ \frac{\lambda}{2} ||B||^{2}_{2} ,}
where Y is a response data matrix, X contains all explanatory variables,
U denotes the score matrix, V is the loading matrix, and B is the effect size matrix.
}
\examples{
library(lfmm)

## sample data
K <- 3
dat <- lfmm_sampler(n = 100, p = 1000, K = K,
                    outlier.prop = 0.1,
                    cs = c(0.8),
                    sigma = 0.2,
                    B.sd = 1.0,
                    U.sd = 1.0,
                    V.sd = 1.0)
## fit an LFMM with K = 3 latent factors
lfmm.res <- lfmm_ridge(Y = dat$Y, X = dat$X, K = 3, lambda = 1e-5)

## plot the effect size matrix
id <- seq_along(lfmm.res$B)
cols <- c('red', 'green')[as.numeric(id \%in\% dat$outlier) + 1]
plot(id, lfmm.res$B, col = cols)
}
\author{
cayek
}
