% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfmm.R
\name{lfmm_ridge_CV}
\alias{lfmm_ridge_CV}
\title{K-flod cross validation of LFMM least-squares estimates with ridge penalty}
\usage{
lfmm_ridge_CV(Y, X, n.fold.row, n.fold.col, lambdas, Ks)
}
\arguments{
\item{Y}{a response variable matrix with n rows and p columns.
Each column corresponds to a distinct response variable (e.g., SNP genotype,
gene expression level, beta-normalized methylation profile, etc).
Response variables must be encoded as numeric.}

\item{X}{an explanatory variable matrix with n rows and d columns.
Each column corresponds to a distinct explanatory variable (eg. phenotype).
Explanatory variables must be encoded as numeric.}

\item{n.fold.row}{number of folds along rows.}

\item{lambdas}{a list of numeric values for the regularization parameter.}

\item{Ks}{a list of integer for the number of latent factors in the regression model.}

\item{p.fold.col}{number of folds along columns.}
}
\value{
a dataframe with prediction error for each value of lambda and K
}
\description{
This function split the dataset into a train set and test set to compute a
prediction error. The function \code{\link{lfmm_ridge}} is run with the
train set and the prediction error is evaluated with the test set.
}
\examples{
library(ggplot2)
library(lfmm)

 ## sample data
 K <- 3
 dat <- lfmm_sampler(n = 100, p = 1000, K = K,
                     outlier.prop = 0.1,
                     cs = c(0.8),
                     sigma = 0.2,
                     B.sd = 1.0,
                     U.sd = 1.0,
                     V.sd = 1.0)

 ## run cross validation
 errs <- lfmm_ridge_CV(Y = dat$Y,
                         X = dat$X,
                         n.fold.row = 5,
                         n.fold.col = 5,
                         lambdas = c(1e-10, 1, 1e20),
                         Ks = c(1,2,3,4,5,6))

 ## plot error
 ggplot(errs, aes(y = err, x = as.factor(K))) +
   geom_boxplot() +
   facet_grid(lambda ~ ., scale = "free")

 ggplot(errs, aes(y = err, x = as.factor(lambda))) +
   geom_boxplot() +
   facet_grid(K ~ ., scales = "free")

}
\seealso{
\code{\link{lfmm_ridge}}
}
\author{
cayek
}
