% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{lfmm_sampler}
\alias{lfmm_sampler}
\title{LFMM generative data sampler}
\usage{
lfmm_sampler(n, p, K, outlier.prop, cs, sigma = 0.2, B.sd = 1, B.mean = 0,
  U.sd = 1, V.sd = 1)
}
\arguments{
\item{n}{number of observations.}

\item{p}{number of output variables.}

\item{K}{number of latent variables.}

\item{outlier.prop}{proportion of outlier.}

\item{cs}{correlation with between X and U.}

\item{sigma}{standard deviation of the error.}

\item{B.sd}{standard deviation of B.}

\item{B.mean}{mean of B.}

\item{U.sd}{standard deviation of U.}

\item{V.sd}{standard deviation of V.}
}
\value{
A list with sampled data.
}
\description{
Sample data with the latent factor model.
}
\details{
\code{lfmm_sample()} sample an output matrix Y and primary matrix X such that

Y = U t(V) + X t(B) + Epsilon.

U,V, B and Epsilon are sampled with normal multivariate distribution.
Moreover U and X are such that \code{cor(U[,i], X) = cs[i]}.
}
\examples{

dat <- lfmm_sampler(n = 100, p = 1000, K = 3,
                    outlier.prop = 0.1,
                    cs = c(0.8),
                    sigma = 0.2,
                    B.sd = 1.0, B.mean = 0.0,
                    U.sd = 1.0, V.sd = 1.0)
}
\author{
kevin caye
}
