% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfmm.R
\name{lfmm_test}
\alias{lfmm_test}
\title{Hypothesis testing of association of Y with X with correction by latent factor.
This function compute the pvalue of the association test of each column of Y
with X. The hypothesis testing take into account latent variables computed
by lfmm_lasso or lfmm_ridge.}
\usage{
lfmm_test(Y, X, lfmm, calibrate = c("gif", "median+MAD"))
}
\arguments{
\item{Y}{Explained variables matrix. Each column is a variable.}

\item{X}{Explaining variables matrix. Each column is a variable.}

\item{lfmm}{Object returned by lfmm_lasso or lfmm_ridge}

\item{calibrate}{If "gif", pvalue are calibrated by computing the genomic
inflation factor of the zscore. If "median+MAD", pvalue are calibrated by
computing the median and MAD of the zscore. If NULL, pvalue are not
calibrated.}
}
\value{
A list with pvalue, zscore and effect.
}
\description{
Hypothesis testing of association of Y with X with correction by latent factor.
This function compute the pvalue of the association test of each column of Y
with X. The hypothesis testing take into account latent variables computed
by lfmm_lasso or lfmm_ridge.
}
\examples{
library(lfmm)

K <- 3
dat <- lfmm_sampler(n = 100, p = 1000, K = K,
                    outlier.prop = 0.1,
                    cs = c(0.8),
                    sigma = 0.2,
                    B.sd = 1.0,
                    U.sd = 1.0,
                    V.sd = 1.0)
## lfmm
lfmm.res <- lfmm_ridge(Y = dat$Y, X = dat$X, K = 3, lambda = 1e-5)

## hp
hp.res <- lfmm_test(Y = dat$Y, X = dat$X, lfmm = lfmm.res, calibrate = TRUE)

## plot score
id <- seq_along(hp.res$calibrated.score)
cols <- c('red', 'green')[as.numeric(id \%in\% dat$outlier) + 1]
plot(id, hp.res$calibrated.score, col = cols)

## plot pvalue
id <- seq_along(hp.res$calibrated.pvalue)
cols <- c('red', 'green')[as.numeric(id \%in\% dat$outlier) + 1]
plot(id, -log10(hp.res$calibrated.pvalue), col = cols)
}
\author{
cayek
}
