% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix-factorizationR.R
\docType{package}
\name{MatrixFactorizationR}
\alias{MatrixFactorizationR}
\alias{MatrixFactorizationR-package}
\title{R package with matrix factorization algorithms}
\description{
R package with matrix factorization algorithms
}
