% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix-factorizationR.R
\name{MatrixFactorizationR_impute}
\alias{MatrixFactorizationR_impute}
\title{Impute Y with a fitted model.}
\usage{
MatrixFactorizationR_impute(m, dat, ...)
}
\description{
Impute Y with a fitted model.
}
