% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix-factorizationR.R
\name{MatrixFactorizationR_residual_error2}
\alias{MatrixFactorizationR_residual_error2}
\title{Compute the residual error}
\usage{
MatrixFactorizationR_residual_error2(m, dat, ...)
}
\description{
Compute the residual error
}
