% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_input}
\alias{read_input}
\title{.. content for \description{} (no empty lines) ..}
\usage{
read_input(X)
}
\arguments{
\item{X}{input}
}
\value{
output
}
\description{
.. content for \details{} ..
}
\author{
cayek
}
