% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfmm.R
\name{ridge_lfmm}
\alias{ridge_lfmm}
\title{LFMM $L_2$ regularized estimator}
\usage{
ridge_lfmm(Y, X, K, lambda = 1e-05)
}
\arguments{
\item{Y}{Explained variables matrix. Each column is a variable.}

\item{X}{Explaining variables matrix. Each column is a variable.}

\item{K}{Number of latent factor.}

\item{lambda}{The value of regularization term.}
}
\value{
A lfmm class object with following attribute:
\itemize{
\item U the latent variable score matrix.
\item V the latent variable axes matrix.
\item B the effect size matrix.
}
}
\description{
This function compute the $L_2$ regularized least squares estimator of LFMM.
The algorithm optimize the loss function
}
\details{
$$ Lridge(\U, \V, \B) =
\frac{1}{2} \norm{Y - U V^{T} - X B^T}\emph{{F}^2 + \frac{\lambda}{2}
\norm{B}^{2}}{2} $$.
}
\examples{
library(lfmm)

## sample data
K <- 3
dat <- lfmm_sampler(n = 100, p = 1000, K = K,
                    outlier.prop = 0.1,
                    cs = c(0.8),
                    sigma = 0.2,
                    B.sd = 1.0,
                    U.sd = 1.0,
                    V.sd = 1.0)
## run lfmm
lfmm.res <- ridge_lfmm(Y = dat$Y, X = dat$X, K = 3, lambda = 1e-5)

## plot size effect matrix
id <- seq_along(lfmm.res$B)
cols <- c('red', 'green')[as.numeric(id \%in\% dat$outlier) + 1]
plot(id, lfmm.res$B, col = cols)
}
\author{
cayek
}
