% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skin_exposure.R
\docType{data}
\name{skin.exposure}
\alias{skin.exposure}
\title{Simulated (and real) methylation levels for sun exposed patient patients}
\format{A list with 6 arguments: beta.value, phenotype, causal.set, chrpos}
\usage{
data("skin.exposure")
}
\description{
A data set containing normalized beta values, and sun exposure and simulated
phenotypic data for 78 tissue samples.
}
\details{
The variables are:

\itemize{
\item beta.value: 1496 filtered normalized beta values (methyation probabilities)
for 78 tissue samples.
\item exposure: Sun exposure levels for 78 tissue samples.
\item phenotype: Simulated binary phenotypic data for 78 tissue samples.
\item  age: age of patients.
\item gender: sex of patients.
\item tissue: category for tissue samples.
}

Reference: to be filled
}
\keyword{datasets}
