% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfmm.R
\name{effect_size}
\alias{effect_size}
\title{Indirect effect sizes for latent factor models}
\usage{
effect_size(Y, X, object)
}
\arguments{
\item{Y}{a response variable matrix with n rows and p columns.
Each column is a response variable (numeric).}

\item{X}{an explanatory variable with n rows and d = 1 column (numeric).}

\item{object}{an object of class \code{lfmm} returned by the \link{lfmm_lasso}
or \link{lfmm_ridge} function.}
}
\value{
a vector of length p containing all effect sizes for the regression
of X on the matrix Y
}
\description{
This function returns 'indirect' effect sizes for the regression of X (of dimension 1) on the matrix Y,
as usually computed in genome-wide association studies.
}
\examples{
library(lfmm)

## Simulation of 1000 phenotypes (x)
## Only the 10 last variables have significant effect sizes
u <- matrix(rnorm(300, sd = 1), nrow = 100, ncol = 2)
v <- matrix(rnorm(3000, sd = 2), nrow = 2, ncol = 1000)
b <- matrix(c(rep(0, 990), rep(1000, 10)))
x <- 1000 + y\%*\%b + rnorm(100, sd = 100)

## Compute direct effect sizes using lfmm_ridge
mod <- lfmm_ridge(Y = scale(y, scale = F), X = scale(x, scale = F), K = 2)

## Compute indirect effect sizes using lfmm_ridge estimates
b.estimates <- effect_size(scale(y, scale = F), scale(x, scale = F), mod)

## plot the 20 last effect sizes (true values are 0 and 1000)
plot(b.estimates[981:1000] )
}
\author{
cayek, francoio
}
