% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfmm.R
\name{lfmm_lasso}
\alias{lfmm_lasso}
\title{LFMM least-squares estimates with lasso penalty}
\usage{
lfmm_lasso(Y, X, K, nozero.prop = 0.01, lambda.num = 100,
  lambda.min.ratio = 0.01, lambda = NULL, it.max = 100,
  relative.err.epsilon = 1e-06)
}
\arguments{
\item{Y}{a response variable matrix with n rows and p columns.
Each column is a response variable (e.g., SNP genotype,
gene expression level, beta-normalized methylation profile, etc).
Response variables must be encoded as numeric.}

\item{X}{an explanatory variable matrix with n rows and d columns.
Each column corresponds to a distinct explanatory variable (eg. phenotype).
Explanatory variables must be encoded as numeric.}

\item{K}{an integer for the number of latent factors in the regression model.}

\item{nozero.prop}{a numeric value for the expected proportion of non-zero effect sizes.}

\item{lambda.num}{a numeric value for the number of 'lambda' values (obscure).}

\item{lambda.min.ratio}{(obscure parameter) a numeric value for the smallest \code{lambda} value,
A fraction of \code{lambda.max}, the data derived entry value (i.e. the smallest value for
which all coefficients are zero).}

\item{lambda}{(obscure parameter) Smallest value of \code{lambda}. A fraction of 'lambda.max',
the (data derived) entry value (i.e. the smallest value for which all
coefficients are zero).}

\item{it.max}{an integer value for the number of iterations of the algorithm.}

\item{relative.err.epsilon}{a numeric value for a relative convergence error. Determine
whether the algorithm converges or not.}
}
\value{
an object of class \code{lfmm} with the following attributes:
\itemize{
\item U the latent variable score matrix with dimensions n x K,
\item V the latent variable axes matrix with dimensions p x K,
\item B the effect size matrix with dimensions p x d.
}
}
\description{
This function computes regularized least squares estimates
for latent factor mixed models using a lasso penalty.
}
\details{
The algorithm minimizes the following penalized least-squares criterion

The response variable matrix Y and the explanatory variable are centered.
}
\examples{

library(lfmm)

## a GWAS example with Y = SNPs and X = phenotype
data(example.data)
Y <- example.data$genotype
X <- example.data$phenotype

## Fit an LFMM with 6 factors
mod.lfmm <- lfmm_lasso(Y = Y, 
                       X = X, 
                       K = 6,
                       nozero.prop = 0.01)
                       
## Perform association testing using the fitted model:
pv <- lfmm_test(Y = Y, 
                X = X, 
                lfmm = mod.lfmm, 
                calibrate = "gif")
                
## Manhattan plot with causal loci shown
pvalues <- pv$calibrated.pvalue
plot(-log10(pvalues), 
      pch = 19, cex = .2, 
      col = "grey", xlab = "SNP")
      
points(example.data$causal.set, 
       -log10(pvalues)[example.data$causal.set], 
       type = "h", col = "blue")
       
## An EWAS example with Y = methylation data 
## and X = exposure
Y <- scale(skin.exposure$beta.value)
X <- scale(as.numeric(skin.exposure$exposure))

## Fit an LFMM with 2 latent factors
mod.lfmm <- lfmm_lasso(Y = Y,
                       X = X, 
                       K = 2,
                       nozero.prop = 0.01)
                       
## Perform association testing using the fitted model:
pv <- lfmm_test(Y = Y, 
                X = X,
                lfmm = mod.lfmm, 
                calibrate = "gif")
                
## Manhattan plot with true associations shown
pvalues <- pv$calibrated.pvalue
plot(-log10(pvalues), 
     pch = 19, 
     cex = .3,
     xlab = "Probe",
     col = "grey")
     
causal.set <- seq(11, 1496, by = 80)
points(causal.set, 
       -log10(pvalues)[causal.set], 
       col = "blue")
}
\author{
cayek, francoio
}
