% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfmm.R
\name{predict_lfmm}
\alias{predict_lfmm}
\title{Predict polygenic risk scores from latent factor models}
\usage{
predict_lfmm(Y, X, object, fdr.level = 0.1, newdata = NULL)
}
\arguments{
\item{Y}{a response variable matrix with n rows and p columns.
Each column is a response variable (numeric).}

\item{X}{an explanatory variable with n rows and d = 1 column (numeric).}

\item{object}{an object of class \code{lfmm} returned by the \link{lfmm_lasso}
or \link{lfmm_ridge} function.}

\item{fdr.level}{a numeric value for the FDR level in the lfmm test used to define
candidate variables for predicting new phenotypes.}

\item{newdata}{a matrix with n rows and p' columns, and similar to Y, on which
predictions of X will be based. If NULL, Y is used as newdata.}
}
\value{
a list with the following attributes:
- prediction: a vector of length n containing the predicted values for X. If
newdata = NULL, the fitted values are returned.
- candidates: vector of candidate columns of Y on which the predictions are
built.
}
\description{
This function polygenic risk scores from latent factor models. It uses the
'indirect' effect sizes for the regression of X (a single phenotype) on the matrix Y,
for predicting phenotypes fromnew data.
}
\examples{
library(lfmm)

## Simulation of 1000 genotypes for 100 individuals (y)
u <- matrix(rnorm(300, sd = 1), nrow = 100, ncol = 2) 
v <- matrix(rnorm(3000, sd = 2), nrow = 2, ncol = 1000)
y <- matrix(rbinom(100000, size = 2, 
                  prob = 1/(1 + exp(-0.3*(u\%*\%v 
                  + rnorm(100000, sd = 2))))),
            nrow = 100,
            ncol = 1000)

#PCA of genotypes, 3 main axes of variation (K = 2) 
plot(prcomp(y))
  
## Simulation of 1000 phenotypes (x)
## Only the last 10 genotypes have significant effect sizes (b)
b <- matrix(c(rep(0, 990), rep(6000, 10)))
x <- y\%*\%b + rnorm(100, sd = 100)
x <- scale(x, scale = F)

## Compute direct effect sizes using lfmm_ridge
## Note that centering is important (scale = F).
mod <- lfmm_ridge(Y = scale(y, scale = F), 
              X = x,
              K = 2)
              
x.pred <- predict_lfmm(Y = scale(y, scale = F), 
                       X = x,
                       fdr.level = 0.1, 
                       mod)$pred

##Compare simulated and predicted/fitted phenotypes
plot(x, x.pred)
abline(0,1)
abline(lm(x.pred ~ x), col = 2)
}
\author{
cayek, francoio
}
